Crossflow_abla/Crossflow_cross_atten_batch_mlp_rd_skip/Crossflow/requirements1.txt
openai-clip
scikit-learn
opencv-python
torchdiffeq
beautifulsoup4
open_clip_torch
scikit-image
cython
matplotlib
accelerate==0.12.0
absl-py
ml_collections
einops
wandb
ftfy==6.1.1
transformers==4.23.1
timm
tensorboard